Address some comments from @lazear in PR #1
