+++ mzcore-rs/src/ms/mod.rs
@@ -1,3 +1,8 @@
 pub mod spectrum;
 pub mod processing;
 pub mod utils;
+
+pub enum MassType {
+   Monoisotopic,
+   Average
+}
